simpler timeouts based on milliseconds

+++ coroo.h
@@ -10,5 +10,5 @@ CorooThread *coroo_thread_start(size_t stack_size,
 		void *thread_argument);
 void coroo_thread_exit();
 
-short coroo_poll_simple(int fd, short events, int timeout);
-void coroo_poll(struct pollfd *fds, nfds_t nfds, int timeout);
+short coroo_poll_simple(int fd, short events, int64_t timeout);
+void coroo_poll(struct pollfd *fds, nfds_t nfds, int64_t timeout);
